backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
PyJWT==2.8.0
bcrypt==4.1.2
opencv-python==4.9.0.80
numpy==1.26.3
ultralytics==8.1.0
scipy==1.11.4
python-dotenv==1.0.0
